requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-ai-projects==1.0.0b1
pandas==2.2.0 